Set base device serial from HAT eeprom
Recent RevPi base devices are requipped with a HAT eeprom which contains
overlay configuration, serial number and other device specific meta data.
The serial number can be used to set the correct serial in piControl, like
it is already done for the extension modules. If not HAT eeprom is present
or the data is invalid a fallback serial of 1 is used. This is the same
serial number as it was hardcoded before.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>