feat(deb): add permissions for piControl dev
"/dev/piControl0" currently has permissions "0666". This allows all
users of the system to read and write the process image which is bad.
Instead, "/dev/piControl0" should have permissions "0660" and be owned
by the group "picontrol". This way who can read and write the process
image is controlled through the "picontrol" group.

To achieve this a udev rule is added. This udev rule assumes the group
"picontrol" exists which is added and removed in the postinst and postrm
scripts respectively.

A problem here is that the piControl kernel module won't be reloaded by
dh_dkms and neither will udev be triggered to reapply its rules. This
means that after installing the picontrol-dkms package:

1. the old piControl module will be running
2. the udev rules won't have applied to "/dev/piControl0"

This can be fixed by rebooting or reloading the module with

    rmmod piControl && modprobe piControl

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>