build(deb)!: Replace the picontrol-dkms debianization
The new debianization can create binary modules packages. This means no
compiler is needed on the target systems. The `piControl` kernel module
will be built for a single kernel package version. If a new kernel
package is released, the piControl packages need to be rebuilt for this
kernel package.

A detailed description can be found in ./debian/README.Debian.

The picontrol-dkms package is removed for now. It might get added later
again.

Signed-off-by: Philipp Rosenberger <p.rosenberger@kunbus.com>