piControl: compact: reimplement sleep of cycle timer
Currently the cycle timer is implemented by putting the polling process
to (uninterruptible) sleep and waiting for the timer to wake it up again.
However for some reason the sleeping process is woken up _before_ the timer
expires, which results in a "premature cycle" warning in the kernel log
file.

Avoid those premature wakeups by using a waitqueue and testing on a flag
that is explicitly set by the timer routine (note that the used function
wait_event() handles the case of spurious wakeups by checking for the flag
after the process has been woken up and putting it back to sleep if the
flag has not been set).

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>