reuse: make piControl REUSE compliant
REUSE is a specification by the Free Software Foundation Europe (FSFE)
that "defines a standardized method for declaring copyright". The Linux
kernel at least partially supports the use of REUSE and checkpatch.pl
also supports linting REUSE annotations in files.

piControl is a central part of the Revolution Pi software ecosystem and
should thus have clear license information. This is already given for
the most part but is not machine-readable. Making it REUSE compliant
changes this.

In order to be REUSE compliant each file needs to clearly state who the
copyright holder and what the license is. This is achieved here.

Copyright years have been taken from the git log which lists commits
until January 2017.

Link: https://reuse.software/spec-3.0/
Link: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/Documentation/process/license-rules.rst#n423
Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>