build: Move Makefile from src/ to root
Generally the Makefile for a project should be in the top-level
directory of the project as this is the standard location for it. Having
the Makefile in the "src/" directory is confusing because the user first
has to change to the "src/" directory to build the project. Compared to
other projects this is confusing as the Makefile is usually located in
the root of the project.

Move the Makefile from "/src/Makefile" to "/Makefile" to cause less
confusion when building the project.

The ability to build a DKMS package is not impaired and this has been
tested both for building the package and installing it on a host.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>