piControl: remove unnecessary substates from state machine
The state machine in PiBridgeMaster_Run() uses the bool bEntering_s in
various states to differentiate between the first and consecutive enterings
of a state. However some states are only visited once and in this case the
use of bEntering_s does not make any sense and only generates unnecessary
code paths. Remove the use of bEntering_s in those states entirely and thus
simplify the code at these places a bit.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>