piControl: Make gate log messages more useful
If a gateway has not been found via RS485, but gate traffic is detected
on the ethernet lines a fallback mechanism is activated. Unfortunately
the log output is not very clear to the user. Thus reword the messages
to make it more obvious, that this is a fallback mechanism with config
lookup.

Also reword the spurious message "no process image synchronization" to
something more useful.

While at it ensure that all gate related messages use the same prefix.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>