piControl: avoid buffer overflow
In case that pwm data is sent the send buffer has to be 2 bytes larger
than DIO_OUTPUT_DATA_LEN. This is since beside the output data stored in
the process image two additional bytes are used for the channel mask.

Fix this by introducing another define for the size of struct pwm_data
and use this to allocated the send buffer.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>