Do not return positive value in piControlInit()
As a module_init function piControlInit() must not return positive values.
The only allowed values are 0 for success or a negative value in case of
an error. So make sure that all invalid (i.e. positive) values returned by
called functions to indicate an error are mapped to negative error values.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>