Ignore failure of piConfigParse()
With commit 2da3676616fe ("Evaluate return value of piConfigParse()") the
return value of piConfigParse() is not longer ignored. This leads to the
undesired behaviour that loading the piControl module fails in case of an
error in the configuration file.

Revert this change and thus restore the old behaviour in which piControl
loads even if the configuration file is incorrect.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>