Do not try to stop kernel thread that does not exist
In revpi_gate_fini() do not try to stop the revpi_gate_rcv_thread if its
creation in revpi_gate_init() has failed before (this may e.g. happen if
a signal is pending in the current thread). Otherwise kthread_stop is
called on an invalid pointer which results in an invalid address access.

Instead check if the thread actually exists and only then stop it.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>