In piControlOpen() do not reset piControl
In piControlOpen() there is no reason to reset the state machine if it is
not in state running: Either a firmware update is in progress and resetting
could interrupt the update process and leave the IO-module bricked. Or the
startup of piControl has failed for some reason and piControl is not
runnable in general.

So remove the reset completely and instead return an error value to
indicate that the device is (temporarily) unavailable.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>