compact: in cycle timer use completion instead of work queue
Currently the cycle timer is executed in hardware interrupt context. This
leads to occasional "scheduling while atomic" bug messages in the kernel
log.

The reason for theses messages is the call of function wake_up() in the
timer routine, since internally wake_up() uses spinlocks to protect data
structures. In a non-RT kernel the use of spinlocks is perfectly fine in
the timer routine, since spinlocks do not sleep in those kernels. However
the RT version of a spinlock is implemented by means of mutexes which do
sleep and thus must not be used in a hardware interrupt context.

To fix this issue reimplement the cycle timer to use a completion to wait
on instead of a workqueue. The completion function complete() which now
replaces wake_up() can be safely called even from hardware interrupt
context.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>