Remove unused counter for open instances
The global counter PnAppCon tracks the number of open '/dev/piControl0'
device file instances. However since setting the counter is done without
any protection it is vulnerable for race conditions if multiple threads
open the device file at the same time.

Furthermore the counter is not used except for debugging output in some
macros that are deactivated anyway. So get rid of the whole instance
counting implementation.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>