Retry if flash erase during firmware update failed
The call of function fwuEraseFlash() as part of a firmware update may fail
although the flash erase has actually been executed on the RevPi IO-Module.
This can happen if the response to a flash erase request gets corrupted on
the pibridge after a successful erase.

Since aborting the firmware update in this case leads to a non-functional
IO-module, repeat the erase, even if that means that the module firmware
will execute the erase multiple times.
Try at most 50 times and only abort the firmware update task if all
attempts failed.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>