piControl: remove unused variable "done"
The variable "done" in the functions "PiBridgeMater_Adjust" and
"revpi_compact_adjust_config" is supposed to signal when the inner loop
is done. This is not needed as all assignments to "done" inside the loop
are followed by a "break", exiting the loop anyways. The variable isn't
used anywhere else in the functions.

Remove the unused "done" variable altogether.

Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>