compact: fix race condition in cycle timer
In cycletimer_sleep() the completion is initialized after the timer is
started. This opens a race condition in which the timer may expire and
fire before the completion is even initialized. As a result the task
waiting for the completion hangs forever.
Fix this by initializing the completion before the timer is started.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>