Set correct return value in do_tree()
With commit 33e8673811f5 ("piConfig: free alloc'd resources in do_tree on
err") the error handling was refactored. As a consequence if no error
occurred the last value stored in 'ret' is the return value of
json_parser_is_done() which is '1'.

Since 'ret' is returned to the caller which expects 0 at no error, the
caller interprets the successful function call of json_parser_is_done() as
a failure, leading to a premature abortion of the device detection
handling.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>