Fix flaw in parser function
In function do_tree() the pointer 'root_structure' is not initialized if
parsing the configuration file failed (e.g because the file does not
exist).

Return an error in this case, since otherwise the caller (piConfigParse())
proceeds with calling find_devices() which eventually tries to dereference
'root_structure' which then results in a kernel splat due to a NULL pointer
access.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>