Reduce number of attempts to erase the flash
With 50 the current number of attempts to erase the flash for a firmware
update is extremely high. If there are so many errors during the piBridge
communication that flash erase only succeeds at the 50th attempt, it better
should not have done in the first place, since the following attempts to
flash the new firmware are very likely to fail also under such conditions.
This is bad, since the modules are left bricked in this case. So better
abort firmware flash erase and thus the entire firmware update completely
if the command to erase the flash does not succeed after 5 attempts.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>