Fix io cycle interval duration
The io cycle is supposed to have a fix interval duration. However the
current implementation with ktime_add_ns() and hrtimer_start() does not
work as expected:
Instead of the desired 5msec a cycle only lasts 1-3 msecs (in a setup with
one DIO connected to a connect4). Also the interval time varies between 1
and 3 msecs.
So instead use a combination of hrtimer_forward() and hrtimer_restart()
which results in the desired fixed interval.

Also omit to extend the time interval in case that a cycle take longer.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>