Do not use different calculations for cycle time
In the io-thread the cycle time is calculated by means of timestamps taken
with hrtimer_cb_get_time() to store it in the process image. Another cycle
time calculation which is based on get_time() is used for a log output at
module unload.
Use for both cases the same calculation based on hrtimer_cb_get_time() and
get rid of the get_time() based calculation.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>