Make cycle attributes accessible via sysfs
Make the cycle time configurable via sysfs. Allow a max of 45 msec and use
the default of 5 msecs as before. Add further sysfs entries that allow to
read the last, min and max cycle times.
Implement a reset of the min and max cycle time if (only) a carriage return
is written to the respective sysfs file.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>