piControl: allow reset ioctl while bridge is stopped
Currently piControlOpen() rejects open() if the PiBridge state machine
is not running. This makes it impossible to recover from a STOP state
such as too many RS485 errors without unloading and reloading the
module.

Drop the waitRunning() check in piControlOpen() so the device can be
opened while stopped. Restrict ioctls in this case to reset and config
operations by extending the !isRunning() check to also allow KB_RESET.

This allows userspace to issue a reset ioctl to restart the state
machine after a STOP without reloading the driver.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>