Fix locking for io telegrams
The process executing an ioctl already uses the lockUserTel
mutex to access the io telegram data. Do the same in the kernel thread
that sends the telegrams, to ensure the data integrity.

To avoid a deadlock in which the ioctl executing process waits for the
result while holding the mutex which is required by the kernel thread,
release the mutex temporarily before waiting for the result and reclaim
it as soon as the results are available.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>