Fix locking for gate telegrams
The process executing an KB_CONFIG_SEND ioctl already uses the lockGateTel
mutex to access the gate telegram data. Do the same in the kernel thread
that sends the telegrams, to ensure the data integrity.

To avoid a deadlock in which the ioctl executing process waits for the
result while holding the mutex which is required by the kernel thread,
release the mutex temporarily before waiting for the result and reclaim
it as soon as the results are available.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>