Allow to send gate/io telegrams in any state
Currently a user specified gate telegram can only be sent if piControl is
in state piBridgeStop, while a user specified io telegram can only be sent
if piControl is in the state of io-communication.
Allow both type of telegrams to be sent in both states to achieve the
highest level of flexibility.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>