Implement cycle timer with fixed time grid
With commit a3a234b9087a ("Fix io cycle interval duration") the erroneous
implementation of the cycle timer was corrected to provide a fix cycle
interval.
However this solution uses relative offsets to set the next timer
expiration time which may result in slight deviations from a fixed time
grid or even bigger shifts in case of missed cycles. If for example a cycle
duration of 5 ms is specified the expiration times may be (starting from
time 0): 0, 5, 10, 15 in the best case. But they may also be 0, 4, 11, 16
in case that the timer did not expire exactly at its last calculated
expiration time.
And in case of missed cycles it may even be 0, 4, 11, 16, (overlong cycle
of 8 msec) 24, 29, ...

So use an absolute expiration time instead and calculate the next time to
expire as an offset from the last expiration time. With this the calculated
expiration times for 5 ms cycles are (starting from time 0): 0, 5, 10, 15,
etc. In case of missed cycles it may be 0, 5, 10, 15, (overlong cycle) 25,
30, ...

Note that with the new cycle timer implementation (hrtimer_forward_now() +
hrtimer_start_expires() instead of hrtimer_forward() + hrtimer_restart())
the minimum time span between timer start and notification of the waiting
thread in the timer callback and thus the minimum time span in which no
bridge-state lock is held during a cycle has been drastically reduced (from
about 500 usecs to only a few usecs on a connect 4).
This makes it hard for processes that require the bridge-state lock, too,
e.g. when checking the running state of the pibridge from userspace.

So provide a minimal time span of 500 usecs in which the lock is not held
by never setting the timer expiration time to a lower value.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>