Implement firmware rescue feature
The firmware bootloader checks the firmware binary CRC before it starts up
the firmware. If this CRC is incorrect, e.g due to a prematurely aborted
firmware update, the firmware is not started. The bootloader switches into
a 'rescue' mode instead in which it accepts gateway datagrams which can be
used to flash a new firmware.

In this case the RevPi master module does not recognize the module in the
initial enumeration phase and the module is considered as not being
connected to the PiBridge.

Support rescue mode by allowing userspace to set the
PICONTROL_FIRMWARE_RESCUE_MODE as well as the modules hardware revision for
a firmware update. Skip the check for the connection state of a module and
try to upload the firmware nevertheless.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>