Use flag to check the PiBridge running state
Introduce the new flag PICONTROL_DEV_FLAG_RUNNING and set it to track the
'running' state of the PiBridge.
In function isRunning() check for this flag instead for the eBridgeState
enum which is protected by the lockBridgeState.

This prevents the callers from having to wait until the lockBridgeState is
available and thus accelerates all accesses from userspace that check for
the running state.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>