compact: do not forget sentinel in lookup table
The gpiod_lookup_table requires a sentinel (empty list entry) at the end
of the table. This is used by functions like gpiod_get_array() to determine
the last valid table entry and omitting it may result in a kernel oops due
to an invalid pointer access.

So add the missing NULL entry to avoid this issue.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>