Register platform device driver also for Flat and Compact
Beside for the RevPi Core/Connect register piControl also for Compact and
Flat as a platform device driver.
This allows to defer the piControl initialization (via the deferred probe()
mechanism) in case that one or more devices required by the Compact or
Flat implementation are not yet available at boot time.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>