avoid piControl_init_sysfs clean up before use
The "piControl_init_sysfs" function is marked with "__init" which
indicated it can be cleaned up after initialisation. The calling
function "pibridge_probe" is *not* an initialization function. This can
lead to "pibridge_probe" trying to call "piControl_init_sysfs", which is
already cleaned up and inaccessible.

Remove the "__init" marker from "piControl_init_sysfs" to avoid this.

Fixes: 43b028beda43 ("Register platform device driver also for Flat and Compact")
Signed-off-by: Thomas Böhler <t.boehler@kunbus.com>