Map internal message errors to -EIO
There are multiple reasons why sending an internal telegram can fail. Map
all possible errors to -EIO so that userspace is not confronted with the
internals of telegram transmission and reception.
Print a message with the original error into the kernel log, so that this
information is still available if needed.

Signed-off-by: Lino Sanfilippo <l.sanfilippo@kunbus.com>